Refactor JSON resource generation
* Upgrade serde to 0.9, to enable the untagged enum feature of Serde. We
use it so that we do not have to hardcode our id as a String type.
Any type that's convertible to a JsonApiId can be used as the id type of
a given resource struct.
* use Into<JsonApiId> to convert from types that can be used as a
JsonApiId
* Instead of using macros to generate the desired JSON, move as much as
possible to the jsonapi crate proper instead of having the bulk of the
logic in jsonapi-derive.