WIP refactoring of query string parsing
Move as much of the query string parsing logic as possible to jsonapi
proper, instead of having it in the jsonapi-derive crate.

* Add a QueryString<'a> trait, which is implemented when using the
JsonApi macro. This sets the SortField and FilterField structs on it
per serializable type.
* Add a TryFrom trait, which is intended to work as the TryFrom crate in
Rust proper. This trait is still unstable, so import it verbatim for
now. This is used to parse the parameters from a query string.