Extract some of the JSON conversion from macro
Move the From impl, which does the conversion from a type T and a Params
type, from the jsonapi-derive crate to jsonapi proper. The only thing
regarding the conversion that's still in the macro is the part that
converts to the JSONAPI attributes for a given type. This part cannot be
removed sadly.