Refactor request traits
Attempting to use the repository pattern in rust doesn't scale too well,
so instead of having a JsonApiService for all HTTP methods, use one
trait per HTTP method that a struct will implement instead.

The structs that a user should implement directly are JsonGet, JsonIndex
and JsonDelete (and later on JsonPatch and JsonPost). If any of these
traits are implemented, a corresponding From[Get|Index|Post|Patch|Delete]
is also automatically implemented, which handles the conversion from a
given struct to its JSONAPI representation, as well as a
[Get|Post|Patch|Index|Delete]Handler, which handles the Iron request.