Add an Option deserializer
When deserializing we need to be able to differentiate when a property
has been explicitly set to null and when a property has not been sent at
all. Since serde does not provide this by default, we need to add our
own custom deserializer.

This commit fixes the failing patch test; now that the deserializer is
in place, we can finally do patches of JSONAPI documents.