Use a router builder instead of extension methods
Remove all the extension methods that were previously used to setup the
routes for the Json* types. Instead we'll use a wrapper type around the
Router, since partially the code is much simpler and partially we need
to setup the body parser, and it's simpler to have the setup in the
router builder than it is for a user to have to set that up on their
own.