Reduce the cloning required when generating code
Previously we have been forced to clone every ident when generating
tokens from them. There's actually no need to do so, we just have to be
sure to send along the idents after extracting them.

We ensure that we're only dealing with structs with named fields by
ensuring that we match on `VariantData::Struct` in `get_attrs_and_id`.
Thereafter, we extract all ident fields and put them into a JsonApiField
struct, which we then can use without having to clone each field.