Reduce iterations and cloning in code generation
Where possible reduce the number of iterations by running all token
generation in the same loop, instead of doing a collect for each
individual list.

The remaining clones have also been removed, except for the cloning of
the ident in `get_attrs_and_id`. Remove the Serialize and Deserialize
traits for the Builder type, since there's no need for the builder to be
able to (de)serialize.