Ids should always be strings
According to the JSONAPI spec, an id should always be a string. Since we
always want to represent the ids as strings in the JSON representation,
there's no point in having the `JsonApiId` type. This doesn't affect any
client code, since like before we still allow any id type that
implements `FromStr`.