Add `TryFrom` implementation
This commit adds a `TryFrom` implementation for any type that implements
`ToJson`. This converts a `JsonApiData<T::Attrs>` to a `T`, and means
that we can get rid of the manual boilerplate converting a
`JsonApiData<T::Attrs>` to a `T`.

In order to do this, we now enforce that any type implementing
`JsonApiBuilder<T>` also has to implement `Default`.

This also moves the builder logic in rustiful-derive to its own
module, instead of having it within the `json` module.