Fix for Travis failure
This fixes an error in Travis where converting from a `Self::Resource`
to a `Self` will set a client supplied id to a default value, instead of
the actual id.

This commit adds a check to see whether an id has been set on the json
resource, and takes the id into account when doing the conversion.