Return JsonApiData instead of Self
Different variations of `Self` are currently returned in the various
resource traits (JsonGet, JsonPatch etc). It is way less magical to
return JsonApiData, not least since it will be easier for a user to
customize the returned output, and it's way more clear what is actually
returned. It does get a bit awkward to use the `Into` trait directly,
especially for lists, which is why I've added an `IntoJson` trait. This
takes `self` and a parameter reference to make it easy to convert to a
JsonApiData type.