Add params for JsonPost and JsonPatch
The JSONAPI spec states that at the very least there should be support
for sparse fieldsets when creating or updating a resource. Therefore,
let's add parameters for JsonPost and JsonPatch.

This also gives the added bonus of a more consistent API, since after
to JsonApiData. Having some cases where we need to send in a default
value and some cases where we use the parameter that's sent in as an
argument is a bit inconsistent.