Return Vec<JsonApiData<T>> in JsonApiArray::data
We always want to return Vec<JsonApiData<T>> in the data field. We
already do this in practice, but to make it more clear state that data
will always return `Vec<JsonApiData<T>>`. This makes it a bit less wordy
when there's a need to declare the type of something that is a
`JsonApiArray<T>`.