Reverse rustiful-derive dependency
There's no good reason for rustiful-derive to depend on rustiful. On the
other hand, adding an optional dependency for rustiful-derive on
rustiful makes it easier to test things within the rustiful module.

The `extern crate rustiful as _rustiful` within the derive files enables
a user to derive `JsonApi` without having to type
`extern crate rustiful`.