Simplify Iron handlers
Remove the `From{Method}` traits and default impls, since this
complicates things needlessly. The things that can be reused by handler
implementations are now instead free (generic) functions. The handler
methods are all renamed to `respond` for the sake of consistency since
they are all differentiated by their types anyway, and they all take an
Iron request as an argument.