Handler lifetime and type parameter changes
There's no need for a lifetime on the trait itself; rather we should
have the lifetimes on the `respond` functions. This makes it easier to
do blanket impls on the handlers, since there is no need for the
handlers to have any type parameters.

This also means we can remove the `autoimpl` dependency, since it only
works for traits with type parameters anyway.

This commit also removes any `where` clauses from the handlers which are
not needed, notably `Deserialize` and (some) of the `'static` bounds.