Add patch method on `JsonPatch`
Currently it's quite unclear on how to update a resource. The current
way of doing so would be `(Self, JsonApiData<Self>).try_into()`, but
it's not clear and pretty easy to forget.

Therefore, add a method which wraps this functionality on `JsonPatch`,
to make it clear how a resource should be updated.