Explicitly set status on error
For `FromRequest`, set the HTTP status code on error. This means that
the API changes slightly; instead of returning just the error type we
return a tuple of an error and a HTTP status.

The same thing applies for the handler traits as well; on an error a
tuple of an error and a HTTP status will be returned as well. In order
not to tie the status type to a specific implementation we make the
handler traits extend another trait, `Handler`, that only has one type
(for now): `Status`.

In `iron/mod.rs` there is a blanket impl for all types that implement
`JsonApiResource`, which sets the `Status` type to the `Status` type
found in Iron/Hyper.

With these changes we also can remove `FromRequestError`,
`BodyParserError` and `RepositoryError`, since there's no longer a need
to wrap user-defined errors in these errors.