Run tests sequentially
We need to run the tests that use SQLite sequentially, since otherwise
the tests will error. The tests that depended on executing sequentially
were previously run as an ignored category, which meant we needed to
run the tests twice; first without the ignored tests and then another
test run with the ignored tests.

We can simply just use a mutex for the tests that need to be run
sequentially, which means that there's no risk of not running these
tests for whatever reason.