Refactor JsonApiData type parameter
There's not actually a need to store the `type` parameter as a string in
the JsonApiData model; all we need to know when deserialising is that
the `type` parameter value is valid. During serialisation we will always
serialise the `type` parameter as `<T as ToJson>::TYPE_NAME`. Since
we already have the type info, it's pointless to store the type as a
string.

Also, since there's now support for const parameters in traits, we store
the `TYPE_NAME` as a const instead of having a method for it.